Project_02/requirements.txt
# LangGraph orchestration
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0

# Text extraction
pdfplumber>=0.10.0
python-docx>=1.0.0

# Diffbot API
requests>=2.31.0

# Entity resolution
rapidfuzz>=3.5.0

# Validation
pydantic>=2.5.0

# Graph database
neo4j>=5.15.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.7

# UI
streamlit>=1.35.0
pandas>=2.0.0

# Config
python-dotenv>=1.0.0
